+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/blocks/needbraces/InputNeedBracesSingleLineLambda.java
@@ -1,6 +1,6 @@
-package com.puppycrawl.tools.checkstyle.checks.blocks;
+package com.puppycrawl.tools.checkstyle.checks.blocks.needbraces;
 
-public class InputSingleLineLambda {
+public class InputNeedBracesSingleLineLambda {
     
     static Runnable r1 = ()->String.CASE_INSENSITIVE_ORDER.equals("Hello world one!");
     static Runnable r2 = () -> String.CASE_INSENSITIVE_ORDER.equals("Hello world two!");
