+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/blocks/needbraces/InputNeedBracesSingleLineStatements.java
@@ -1,6 +1,6 @@
-package com.puppycrawl.tools.checkstyle.checks.blocks;
+package com.puppycrawl.tools.checkstyle.checks.blocks.needbraces;
 
-public class InputBracesSingleLineStatements
+public class InputNeedBracesSingleLineStatements
 {
     private static class SomeClass {
         boolean flag = true;
